[improve][client] PIP-478: let TokenAuthenticationV5 parse the v4 token parameters
Preparation for resolving authPluginClassName straight to a v5 plugin: the v5 token body could
only be constructed from a Supplier<String>, so a client configured with
authPluginClassName=...AuthenticationToken had no way to reach it by name.

TokenAuthenticationV5 gains fromParams(String) and the tokenSupplier(String) parser behind it,
accepting exactly the forms the v4 plugin accepted — token:<literal>, file:<uri>, a JSON object
with a token member, or the bare token. The two serializable suppliers move onto it as
LiteralTokenSupplier / FileTokenSupplier.

The v4 AuthenticationToken now delegates its configure(String) to that parser rather than
keeping a second copy, so the by-name path and the v4 shim cannot drift apart.
